requirements.txt
aadict==0.2.3
aiohttp==3.5.4
alembic==1.0.8
apispec==1.3.3
apparmor==2.13.2
apturl==0.5.2
asn1crypto==0.24.0
asset==0.6.12
astral==1.10.1
async-timeout==3.0.1
attrs==19.1.0
Babel==2.6.0
backcall==0.1.0
bcrypt==3.1.6
bleach==3.1.0
blinker==1.4
Brlapi==0.6.7
catkin-pkg==0.4.10
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
command-not-found==0.3
cryptography==2.6.1
cssmin==0.2.0
cupshelpers==1.0
cycler==0.10.0
dash==0.43.0
dash-core-components==0.48.0
dash-daq==0.1.5
dash-html-components==0.16.0
dash-renderer==0.24.0
dash-table==3.7.0
decorator==4.4.0
defer==1.0.6
defusedxml==0.5.0
distro==1.3.0
distro-info===0.21ubuntu2
docutils==0.14
empy==3.3.2
entrypoints==0.3
Flask==1.0.3
Flask-AppBuilder==2.1.4
Flask-Assets==0.12
Flask-Babel==0.12.2
Flask-BabelEx==0.9.3
Flask-Compress==1.4.0
Flask-Gravatar==0.4.2
Flask-HTMLmin==1.3.2
Flask-JWT-Extended==3.18.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
Flask-OpenID==1.2.5
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-SeaSurf==0.2.2
Flask-Security==1.7.5
Flask-Session==0.3.1
Flask-SQLAlchemy==2.4.0
Flask-User==1.0.1.5
Flask-WTF==0.14.2
globre==0.1.5
gunicorn==19.9.0
homeassistant==0.93.2
html5lib==1.0.1
htmlmin==0.1.12
httplib2==0.11.3
idna==2.8
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
jsonschema==3.0.1
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
keyring==17.1.1
keyrings.alt==3.1.1
kiwisolver==1.1.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lessc==0.1.2
LibAppArmor==2.13.2
louis==3.8.0
lxml==4.3.2
macaroonbakery==1.2.1
Mako==1.0.8
MarkupSafe==1.1.1
marshmallow==2.19.3
marshmallow-enum==1.4.1
marshmallow-sqlalchemy==0.16.4
matplotlib==3.1.0
mistune==0.8.4
multidict==4.5.2
nbconvert==5.4.0
nbformat==4.4.0
netifaces==0.10.4
notebook==5.7.4
numpy==1.16.4
oauthlib==2.1.0
olefile==0.46
pandas==0.24.2
pandas-datareader==0.7.0
pandocfilters==1.4.2
paramiko==2.4.2
parso==0.4.0
passlib==1.7.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
plotly==3.10.0
prison==0.1.1
prometheus-client==0.5.0
prompt-toolkit==2.0.9
protobuf==3.6.1
psutil==5.5.1
psycopg2==2.8.2
pyasn1==0.4.2
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
pyexpect==1.0.19
Pygments==2.3.1
PyGObject==3.32.0
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.2.0
pyRFC3339==1.1
pyrsistent==0.14.11
python-apt==1.8.4
python-dateutil==2.8.0
python-debian==0.1.34
python-editor==1.0.4
python-slugify==3.0.2
python3-openid==3.1.0
pytz==2019.1
pyxdg==0.25
PyXiaomiGateway==0.12.3
PyYAML==3.13
pyzmq==17.1.2
redis==3.2.1
reportlab==3.5.18
requests==2.21.0
requests-unixsocket==0.1.5
retrying==1.3.3
roman==2.0.0
ruamel.yaml==0.15.94
SecretStorage==2.3.1
selenium==3.141.0
Send2Trash==1.5.0
simplegeneric==0.8.1
simplejson==3.16.0
six==1.12.0
speaklater==1.3
SQLAlchemy==1.3.4
SQLAlchemy-Utils==0.34.0
sqlparse==0.2.4
ssh-import-id==5.7
sshtunnel==0.1.4
style==1.1.0
system-service==0.3
systemd-python==234
terminado==0.8.1
testpath==0.4.2
text-unidecode==1.2
tornado==6.0.2
traitlets==4.3.2
tzlocal==1.5.1
ubuntu-advantage-tools==19.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
update==0.0.1
urllib3==1.24.3
usb-creator==0.3.3
voluptuous==0.11.5
voluptuous-serialize==2.1.0
wadllib==1.3.3
wcwidth==0.1.7
webassets==0.12.1
webencodings==0.5.1
Werkzeug==0.15.4
wrapt==1.11.1
WTForms==2.2.1
xkit==0.0.0
yarl==1.3.0
